Move swap-domain components to appropriate folder
