+++ src/cow-react/pages/NewSwap/styled.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components/macro'
 import SwapHeader from 'components/swap/SwapHeader'
-import { RemoveRecipient } from 'pages/Swap/components/RemoveRecipient'
+import { RemoveRecipient } from 'cow-react/swap/smart/RemoveRecipient'
 import { MEDIA_WIDTHS } from 'theme'
 
 export const Container = styled.div`
