+++ src/cow-react/swap/dumb/CurrencyArrowSeparator/__fixtures__/CurrencyArrowSeparator.tsx
@@ -1,7 +1,4 @@
-import {
-  CurrencyArrowSeparator,
-  CurrencyArrowSeparatorProps,
-} from 'cow-react/pages/NewSwap/pureComponents/CurrencyArrowSeparator'
+import { CurrencyArrowSeparator, CurrencyArrowSeparatorProps } from 'cow-react/swap/dumb/CurrencyArrowSeparator'
 
 const defaultProps: CurrencyArrowSeparatorProps = {
   isLoading: false,
