+++ src/cow-react/swap/dumb/CurrencyArrowSeparator/index.tsx
