+++ src/cow-react/swap/dumb/CurrencyArrowSeparator/style-mixins.tsx
