+++ src/cow-react/swap/dumb/CurrencyArrowSeparator/styled.tsx
