+++ src/cow-react/swap/dumb/CurrencyInputPanel/index.tsx
@@ -8,7 +8,7 @@ import { CurrencyInfo } from 'cow-react/pages/NewSwap/typings'
 import { FiatValue } from 'components/CurrencyInputPanel/FiatValue'
 import { Trans } from '@lingui/macro'
 import { PriceImpact } from 'hooks/usePriceImpact'
-import { ReceiveAmount } from 'cow-react/pages/NewSwap/pureComponents/ReceiveAmount'
+import { ReceiveAmount } from 'cow-react/swap/dumb/ReceiveAmount'
 import { BalanceAndSubsidy } from 'hooks/useCowBalanceAndSubsidy'
 import { setMaxSellTokensAnalytics } from 'utils/analytics'
 import { SwapActions } from 'state/swap/hooks'
