+++ src/cow-react/swap/dumb/CurrencyInputPanel/styled.tsx
