+++ src/cow-react/swap/dumb/ReceiveAmount/styled.tsx
