+++ src/cow-react/swap/dumb/ReceiveAmountInfo/index.tsx
