+++ src/cow-react/swap/dumb/ReceiveAmountInfo/styled.tsx
