+++ src/cow-react/swap/dumb/SwapButton/SwapButton.tsx
@@ -10,7 +10,7 @@ import { GreyCard } from 'components/Card'
 import { GpEther } from 'constants/tokens'
 import { SupportedChainId } from 'constants/chains'
 import { AutoColumn } from 'components/Column'
-import { ApproveButton, ApproveButtonProps } from 'pages/Swap/components/ApproveButton'
+import { ApproveButton, ApproveButtonProps } from 'cow-react/swap/smart/ApproveButton'
 import * as styledEl from './styled'
 import { HandleSwapCallback } from 'cow-react/swap/hooks/useHandleSwap'
 import { isSwapButtonPropsEqual } from 'cow-react/pages/NewSwap/propsChecker'
