+++ src/cow-react/swap/dumb/TradeRates/index.tsx
@@ -4,13 +4,13 @@ import { InfoIcon } from 'components/InfoIcon'
 import { SUBSIDY_INFO_MESSAGE } from 'components/CowSubsidyModal/constants'
 import { useOpenModal } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
-import { Price } from 'pages/Swap/components/Price'
+import { Price } from 'cow-react/swap/smart/Price'
 import TradeGp from 'state/swap/TradeGp'
 import { INITIAL_ALLOWED_SLIPPAGE_PERCENT } from 'constants/index'
 import { RowSlippage } from 'components/swap/TradeSummary/RowSlippage'
 import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 import { genericPropsChecker } from 'cow-react/pages/NewSwap/propsChecker'
-import { TradeBasicDetails } from 'pages/Swap/components/TradeBasicDetails'
+import { TradeBasicDetails } from 'cow-react/swap/smart/TradeBasicDetails'
 
 const SUBSIDY_INFO_MESSAGE_EXTENDED =
   SUBSIDY_INFO_MESSAGE + '. Click on the discount button on the right for more info.'
