+++ src/cow-react/swap/dumb/modals.tsx
@@ -1,8 +1,8 @@
-import { ConfirmSwapModalSetup, ConfirmSwapModalSetupProps } from 'pages/Swap/components/ConfirmSwapModalSetup'
+import { ConfirmSwapModalSetup, ConfirmSwapModalSetupProps } from 'cow-react/swap/smart/ConfirmSwapModalSetup'
 import EthFlowModal, { EthFlowProps } from 'components/swap/EthFlow'
 import React from 'react'
 import { genericPropsChecker } from 'cow-react/pages/NewSwap/propsChecker'
-import { ImportTokenModal } from 'pages/Swap/components/ImportTokenModal'
+import { ImportTokenModal } from 'cow-react/swap/smart/ImportTokenModal'
 import CowSubsidyModal from 'components/CowSubsidyModal'
 import { useCloseModals } from 'state/application/hooks'
 import { useHistory } from 'react-router-dom'
