+++ src/cow-react/swap/hooks/useSwapButtonContext.ts
@@ -15,9 +15,9 @@ import { swapFlow } from 'cow-react/swap/services/swapFlow'
 import { useApproveCallbackFromTrade } from 'hooks/useApproveCallback'
 import { useGnosisSafeInfo } from 'hooks/useGnosisSafeInfo'
 import { useIsSwapUnsupported } from 'hooks/useIsSwapUnsupported'
-import { ApproveButtonProps } from 'pages/Swap/components/ApproveButton'
+import { ApproveButtonProps } from 'cow-react/swap/smart/ApproveButton'
 import { getSwapButtonState } from 'cow-react/swap/helpers/getSwapButtonState'
-import { SwapButtonContext } from 'pages/Swap/components/SwapButton/SwapButton'
+import { SwapButtonContext } from 'cow-react/swap/dumb/SwapButton/SwapButton'
 import { useGetQuoteAndStatus } from 'state/price/hooks'
 import { OperationType } from 'components/TransactionConfirmationModal'
 import { useTransactionConfirmModal } from 'cow-react/swap/hooks/useTransactionConfirmModal'
