+++ src/cow-react/swap/smart/ApproveButton.tsx
@@ -1,10 +1,10 @@
 import { ButtonConfirmed } from 'components/Button'
 import { ApprovalState, OptionalApproveCallbackParams } from 'hooks/useApproveCallback'
-import { UseERC20PermitState } from '@src/hooks/useERC20Permit'
-import { AutoRow } from '@src/components/Row'
+import { UseERC20PermitState } from 'hooks/useERC20Permit'
+import { AutoRow } from 'components/Row'
 import CurrencyLogo from 'components/CurrencyLogo'
 import { Trans } from '@lingui/macro'
-import Loader from '@src/components/Loader'
+import Loader from 'components/Loader'
 import { CheckCircle, HelpCircle } from 'react-feather'
 import { MouseoverTooltip } from 'components/Tooltip'
 import { ButtonSize } from 'theme'
@@ -16,10 +16,10 @@ import TradeGp from 'state/swap/TradeGp'
 import { Currency, Percent } from '@uniswap/sdk-core'
 import { BigNumber } from '@ethersproject/bignumber'
 import { ThemeContext } from 'styled-components/macro'
-import usePrevious from '@src/hooks/usePrevious'
+import usePrevious from 'hooks/usePrevious'
 import { ButtonError } from 'components/Button'
-import { HandleSwapCallback } from '@src/cow-react/swap/hooks/useHandleSwap'
-import { useSwapConfirmManager } from '@src/cow-react/swap/hooks/useSwapConfirmManager'
+import { HandleSwapCallback } from 'cow-react/swap/hooks/useHandleSwap'
+import { useSwapConfirmManager } from 'cow-react/swap/hooks/useSwapConfirmManager'
 import { TransactionResponse } from '@ethersproject/providers'
 
 export interface ApproveButtonProps {
