+++ src/cow-react/swap/smart/CompatibilityIssuesWarning.tsx
