+++ src/cow-react/swap/smart/ConfirmSwapModalSetup.tsx
@@ -5,8 +5,8 @@ import { Percent } from '@uniswap/sdk-core'
 import { Field } from 'state/swap/actions'
 import { useAtomValue } from 'jotai/utils'
 import { swapConfirmAtom } from 'cow-react/swap/state/swapConfirmAtom'
-import { HandleSwapCallback } from '@src/cow-react/swap/hooks/useHandleSwap'
-import { useSwapConfirmManager } from '@src/cow-react/swap/hooks/useSwapConfirmManager'
+import { HandleSwapCallback } from 'cow-react/swap/hooks/useHandleSwap'
+import { useSwapConfirmManager } from 'cow-react/swap/hooks/useSwapConfirmManager'
 import TransactionConfirmationModal from 'components/TransactionConfirmationModal'
 import { useSwapActionHandlers } from 'state/swap/hooks'
 import { useModalIsOpen } from 'state/application/hooks'
