+++ src/cow-react/swap/smart/TradeBasicDetails.tsx
