+++ src/custom/pages/Swap/SwapMod.tsx
@@ -42,22 +42,22 @@ import { useErrorMessage } from 'hooks/useErrorMessageAndModal'
 import CowSubsidyModal from 'components/CowSubsidyModal'
 import { AlertWrapper } from './styleds'
 import { setMaxSellTokensAnalytics } from 'utils/analytics'
-import { ImportTokenModal } from './components/ImportTokenModal'
-import { CompatibilityIssuesWarning } from './components/CompatibilityIssuesWarning'
-import { ConfirmSwapModalSetup, ConfirmSwapModalSetupProps } from 'pages/Swap/components/ConfirmSwapModalSetup'
+import { ImportTokenModal } from 'cow-react/swap/smart/ImportTokenModal'
+import { CompatibilityIssuesWarning } from 'cow-react/swap/smart/CompatibilityIssuesWarning'
+import { ConfirmSwapModalSetup, ConfirmSwapModalSetupProps } from 'cow-react/swap/smart/ConfirmSwapModalSetup'
 import { useAtomValue } from 'jotai/utils'
 import { swapConfirmAtom } from 'cow-react/swap/state/swapConfirmAtom'
-import { SwapButton, SwapButtonContext } from 'pages/Swap/components/SwapButton/SwapButton'
-import { RemoveRecipient } from 'pages/Swap/components/RemoveRecipient'
-import { Price } from './components/Price'
-import { TradeBasicDetails } from 'pages/Swap/components/TradeBasicDetails'
+import { SwapButton, SwapButtonContext } from 'cow-react/swap/dumb/SwapButton/SwapButton'
+import { RemoveRecipient } from 'cow-react/swap/smart/RemoveRecipient'
+import { Price } from 'cow-react/swap/smart/Price'
+import { TradeBasicDetails } from 'cow-react/swap/smart/TradeBasicDetails'
 import { BottomGrouping } from 'pages/Swap/styled'
 import { ArrowWrapperLoader } from 'components/ArrowWrapperLoader'
 import { HighFeeWarning, NoImpactWarning } from 'components/SwapWarnings'
-import { FeesDiscount } from 'pages/Swap/components/FeesDiscount'
+import { FeesDiscount } from 'cow-react/swap/smart/FeesDiscount'
 import { RouteComponentProps } from 'react-router-dom'
 import EthFlowModal from 'components/swap/EthFlow'
-import { useSwapButtonContext } from '@src/cow-react/swap/hooks/useSwapButtonContext'
+import { useSwapButtonContext } from 'cow-react/swap/hooks/useSwapButtonContext'
 import { Routes } from 'constants/routes'
 
 export default function Swap({ history, location, className }: RouteComponentProps & { className?: string }) {
